# docs/swagger.yaml
basePath: /
definitions:
...
  dto.VehicleTypeCreateRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.VehicleTypeUpdateRequest:
    properties:
      id:
        type: string
      name:
        type: string
    required:
    - id
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
host: localhost:8888
info:
  contact:
    email: mddx56@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestionar motos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Motos API
  version: 0.1.2
paths:
...
  /api/vehicle-types:
    get:
      consumes:
      - application/json
      description: Get a list of all active vehicle types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all vehicle types
      tags:
      - vehicle-types
    post:
      consumes:
      - application/json
      description: Create a new vehicle type with the input payload
      parameters:
      - description: VehicleType Create Request
        in: body
        name: vehicleType
        required: true
        schema:
          $ref: '#/definitions/dto.VehicleTypeCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a new vehicle type
      tags:
      - vehicle-types
  /api/vehicle-types/{id}:
    get:
      consumes:
      - application/json
      description: Get a vehicle type by ID
      parameters:
      - description: VehicleType ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a vehicle type by ID
      tags:
      - vehicle-types
    put:
      consumes:
      - application/json
      description: Update an existing vehicle type with the input payload
      parameters:
      - description: VehicleType ID
        in: path
        name: id
        required: true
        type: string
      - description: VehicleType Update Request
        in: body
        name: vehicleType
        required: true
        schema:
          $ref: '#/definitions/dto.VehicleTypeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update an existing vehicle type
      tags:
      - vehicle-types
  /api/vehicle-types/{id}/status:
    patch:
      consumes:
      - application/json
      description: Soft delete a vehicle type by setting status to false
      parameters:
      - description: VehicleType ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Change status a vehicle type (soft delete)
      tags:
      - vehicle-types
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"